Fixing up Seam Security setup
Looks like this should work now if data is properly setup. Little gotcha
that I had for an hour, Seam Config is not pulled in anywhere, you have
to add it yourself.

Also the Credentials seem like they must be their own class. I'll have
to verify with Shane, but that's what it appears.